# examples/todo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Code generator",
        "version": "0.1"
    },
    "paths": {
        "/api/user/create": {
            "post": {
                "tags": [
                    "---------- User ----------"
                ],
                "summary": "Create User",
                "operationId": "create/user",
                "requestBody": {
                    "description": "Create User object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateUserApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/search": {
            "post": {
                "tags": [
                    "---------- User ----------"
                ],
                "summary": "Search User",
                "operationId": "search/user",
                "requestBody": {
                    "description": "Search User object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchUserApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchUserApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/update": {
            "post": {
                "tags": [
                    "---------- User ----------"
                ],
                "summary": "Update User",
                "operationId": "update/user",
                "requestBody": {
                    "description": "Update User object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateUserApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateUserApiRequest": {
                "title": "CreateUserApiRequest",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateUserApiResponse": {
                "title": "CreateUserApiResponse",
                "properties": {
                    "success": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "hihi": {
                                "type": "string"
                            },
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SearchUserApiRequest": {
                "title": "SearchUserApiRequest",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SearchUserApiResponse": {
                "title": "SearchUserApiResponse",
                "properties": {
                    "success": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/SearchUserResponseItem"
                    }
                },
                "type": "object"
            },
            "SearchUserResponseItem": {
                "title": "SearchUserResponseItem",
                "properties": {
                    "owner": {
                        "$ref": "#/components/schemas/UserDto"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserApiRequest": {
                "title": "UpdateUserApiRequest",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserApiResponse": {
                "title": "UpdateUserApiResponse",
                "properties": {
                    "success": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserDto": {
                "title": "UserDto",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}